Battery Utility: This patch adds the battery utility code to github.
This utility will be used in VMs to fetch the battery related sysfs entry
data from the Host OS to the Guest OS (Android). The Health HAL on the
android will receive the data and take actions accordingly. This will also
help update the battery percentage icon on Android UI.

This version supports the following:
1. Update battery related sysfs entry data to the guest OS.
2. Send battery related data at every 1 second interval.
3. When android is rebooted or connection is lost, then the utility will
again start listening on the port specified for connection.

Signed-off-by: Tanuj Tekriwal <tanuj.tekriwal@intel.com>